+++ internal/mapdecode/CHANGELOG.md
@@ -4,6 +4,8 @@ Releases
 v0.2.0 (unreleased)
 -------------------
 
+-   Added `FieldHook` to intercept values before they are decoded into specific
+    struct fields.
 -   Decode now parses strings if they are found in place of a float, boolean,
     or integer.